+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 hook_update_deploy_tools 7.x-1.x **-**-****
 -----------------------------------------------
+* Issue #2857507: Make Vocabulary::add method.
 
 hook_update_deploy_tools 7.x-1.21 Jan 12, 2017
 -----------------------------------------------